requirements.txt
streamlit==1.41.1
pandas==2.2.3
tabulate==0.9.0
langchain==0.3.16
langchain-community==0.3.16
langchain-neo4j==0.3.0
langchain-groq==0.2.3
langchain-experimental==0.3.4
pypdf==5.2.0
python-dotenv==1.0.1
tiktoken==0.8.0
replicate==1.0.4
json-repair